Use Sinon for mocking.  Fill a test gap for Linux PATH.  Add tests with task runner. (#7054)
