+++ Tasks/UsePythonVersion/Tests/L0.ts
@@ -2,227 +2,55 @@ import * as assert from 'assert';
 import { EOL } from 'os';
 import * as path from 'path';
 
-import * as mockery from 'mockery';
-import * as mockTask from 'vsts-task-lib/mock-task';
+import { MockTestRunner } from 'vsts-task-lib/mock-test';
 
-import { Platform } from '../taskutil';
-import * as usePythonVersion from '../usepythonversion';
-
-/** Reload the unit under test to use mocks that have been registered. */
-function reload(): typeof usePythonVersion {
-    return require('../usepythonversion');
+function didSetVariable(testRunner: MockTestRunner, variableName: string, variableValue: string): boolean {
+    return testRunner.stdOutContained(`##vso[task.setvariable variable=${variableName};issecret=false;]${variableValue}`);
 }
 
 describe('UsePythonVersion L0 Suite', function () {
-    before(function () {
-        mockery.enable({
-            useCleanCache: true,
-            warnOnUnregistered: false
-        });
+    describe('usepythonversion.ts', function () {
+        require('./L0_usepythonversion');
     });
 
-    after(function () {
-        mockery.disable();
+    it('succeeds when version is found', function () {
+        const testFile = path.join(__dirname, 'L0SucceedsWhenVersionIsFound.js');
+        const testRunner = new MockTestRunner(testFile);
+
+        testRunner.run();
+
+        assert(didSetVariable(testRunner, 'pythonLocation', path.join('/', 'Python', '3.6.4', 'x64')));
+        assert.strictEqual(testRunner.stderr.length, 0, 'should not have written to stderr');
+        assert(testRunner.succeeded, 'task should have succeeded');
     });
 
-    afterEach(function () {
-        mockery.deregisterAll();
-        mockery.resetCache();
-    })
+    it('fails when version is not found', function () {
+        const testFile = path.join(__dirname, 'L0FailsWhenVersionIsMissing.js');
+        const testRunner = new MockTestRunner(testFile);
 
-    it('converts Python prerelease versions to the semantic version format', function () {
-        mockery.registerMock('vsts-task-lib/task', mockTask);
-        mockery.registerMock('vsts-task-tool-lib/tool', {});
-        const uut = reload();
+        testRunner.run();
 
-        const testCases = [
-            {
-                versionSpec: '3.x',
-                expected: '3.x'
-            },
-            {
-                versionSpec: '3.3.6',
-                expected: '3.3.6'
-            },
-            {
-                versionSpec: '3.7.0b2',
-                expected: '3.7.0-b2'
-            },
-            {
-                versionSpec: '3.7.0rc',
-                expected: '3.7.0-rc'
-            },
-            {
-                versionSpec: '14.22.100a1000',
-                expected: '14.22.100-a1000'
-            },
-            {
-                versionSpec: '3.6.6b2 || >= 3.7.0rc',
-                expected: '3.6.6-b2 || >= 3.7.0-rc'
-            },
-            {
-                versionSpec: '3.7rc1', // invalid
-                expected: '3.7rc1'
-            },
-        ];
+        const errorMessage = [
+            'loc_mock_VersionNotFound 3.x',
+            'loc_mock_ListAvailableVersions',
+            '2.6.0 (x86)',
+            '2.7.13 (x86)',
+            '2.6.0 (x64)',
+            '2.7.13 (x64)'
+        ].join(EOL);
 
-        for (let tc of testCases) { // Node 5 can't handle destructuring assignment
-            const actual = uut.pythonVersionToSemantic(tc.versionSpec);
-            assert.strictEqual(actual, tc.expected);
-        }
-    })
-
-    it('finds version in cache', async function () {
-        let buildVariables: { [key: string]: string } = {};
-        const mockBuildVariables = {
-            setVariable: (variable: string, value: string) => {
-                buildVariables[variable] = value;
-            },
-            getVariable: (variable: string) => buildVariables[variable]
-        };
-        mockery.registerMock('vsts-task-lib/task', Object.assign({}, mockTask, mockBuildVariables));
-
-        const toolPath = path.join('/', 'Python', '3.6.4', 'x64');
-        mockery.registerMock('vsts-task-tool-lib/tool', {
-            findLocalTool: () => toolPath
-        });
-
-        const uut = reload();
-        const parameters = {
-            versionSpec: '3.6',
-            addToPath: false,
-            architecture: 'x64'
-        };
-
-        assert.strictEqual(buildVariables['pythonLocation'], undefined);
-
-        await uut.usePythonVersion(parameters, Platform.Linux);
-        assert.strictEqual(buildVariables['pythonLocation'], toolPath);
+        assert(testRunner.createdErrorIssue(errorMessage));
+        assert(testRunner.failed, 'task should have failed');
     });
 
-    it('rejects version not in cache', async function (done: MochaDone) {
-        mockery.registerMock('vsts-task-lib/task', mockTask);
-        mockery.registerMock('vsts-task-tool-lib/tool', {
-            findLocalTool: () => null,
-            findLocalToolVersions: () => ['2.6.0', '2.7.13']
-        });
+    it('selects architecture passed as input', function () {
+        const testFile = path.join(__dirname, 'L0SelectsArchitecture.js');
+        const testRunner = new MockTestRunner(testFile);
 
-        const uut = reload();
-        const parameters = {
-            versionSpec: '3.x',
-            addToPath: false,
-            architecture: 'x64'
-        };
+        testRunner.run();
 
-        try {
-            await uut.usePythonVersion(parameters, Platform.Linux);
-            done(new Error('should not have succeeded'));
-        } catch (e) {
-            const expectedMessage = [
-                'loc_mock_VersionNotFound 3.x',
-                'loc_mock_ListAvailableVersions',
-                '2.6.0 (x86)',
-                '2.7.13 (x86)',
-                '2.6.0 (x64)',
-                '2.7.13 (x64)'
-            ].join(EOL);
-
-            assert.strictEqual(e.message, expectedMessage);
-            done();
-        }
-    });
-
-    it('selects architecture passed as input', async function () {
-        let buildVariables: { [key: string]: string } = {};
-        const mockBuildVariables = {
-            setVariable: (variable: string, value: string) => {
-                buildVariables[variable] = value;
-            },
-            getVariable: (variable: string) => buildVariables[variable]
-        };
-        mockery.registerMock('vsts-task-lib/task', Object.assign({}, mockTask, mockBuildVariables));
-
-        const x86ToolPath = path.join('/', 'Python', '3.6.4', 'x86');
-        const x64ToolPath = path.join('/', 'Python', '3.6.4', 'x64');
-        mockery.registerMock('vsts-task-tool-lib/tool', {
-            findLocalTool: (toolName: string, versionSpec: string, arch?: string) => {
-                if (arch === 'x86') {
-                    return x86ToolPath;
-                } else {
-                    return x64ToolPath;
-                }
-            }
-        });
-
-        const uut = reload();
-        const parameters = {
-            versionSpec: '3.6',
-            addToPath: false,
-            architecture: 'x86'
-        };
-
-        assert.strictEqual(buildVariables['pythonLocation'], undefined);
-
-        await uut.usePythonVersion(parameters, Platform.Linux);
-        assert.strictEqual(buildVariables['pythonLocation'], x86ToolPath);
-    });
-
-    it('sets PATH correctly on Linux', async function () {
-        mockery.registerMock('vsts-task-lib/task', mockTask);
-
-        const toolPath = path.join('/', 'Python', '3.6.4', 'x64');
-        mockery.registerMock('vsts-task-tool-lib/tool', {
-            findLocalTool: () => toolPath,
-        });
-
-        let mockPath = '';
-        mockery.registerMock('./toolutil', {
-            prependPathSafe: (s: string) => {
-                mockPath = s + ':' + mockPath;
-            }
-        });
-
-        const uut = reload();
-        const parameters = {
-            versionSpec: '3.6',
-            addToPath: true,
-            architecture: 'x64'
-        };
-
-        await uut.usePythonVersion(parameters, Platform.Linux);
-        assert.strictEqual(`${path.join(toolPath, 'bin')}:${toolPath}:`, mockPath);
-    });
-
-    it('sets PATH correctly on Windows', async function () {
-        mockery.registerMock('vsts-task-lib/task', mockTask);
-
-        const toolPath = path.join('/', 'Python', '3.6.4', 'x64');
-        mockery.registerMock('vsts-task-tool-lib/tool', {
-            findLocalTool: () => toolPath
-        });
-
-        let mockPath = '';
-        mockery.registerMock('./toolutil', {
-            prependPathSafe: (s: string) => {
-                mockPath = s + ';' + mockPath;
-            }
-        });
-
-        process.env['APPDATA'] = '/mock-appdata';
-
-        const uut = reload();
-        const parameters = {
-            versionSpec: '3.6',
-            addToPath: true,
-            architecture: 'x64'
-        };
-
-        await uut.usePythonVersion(parameters, Platform.Windows);
-
-        // On Windows, must add the two "Scripts" directories to PATH as well
-        const expectedScripts = path.join(toolPath, 'Scripts');
-        const expectedUserScripts = path.join(process.env['APPDATA'], 'Python', 'Python36', 'Scripts');
-        const expectedPath = `${expectedUserScripts};${expectedScripts};${toolPath};`;
-        assert.strictEqual(expectedPath, mockPath);
+        assert(didSetVariable(testRunner, 'pythonLocation', 'x86ToolPath'));
+        assert.strictEqual(testRunner.stderr.length, 0, 'should not have written to stderr');
+        assert(testRunner.succeeded, 'task should have succeeded');
     });
 });
