+++ Tasks/UsePythonVersion/Tests/package.json
@@ -19,6 +19,8 @@
   "devDependencies": {
     "@types/mocha": "^2.2.48",
     "@types/mockery": "^1.4.29",
-    "mockery": "^2.1.0"
+    "@types/sinon": "^4.3.1",
+    "mockery": "^2.1.0",
+    "sinon": "^4.5.0"
   }
 }
